// src/features/callTracker/models/CallTrackerModel.ts
import type { ApiResponse } from "@/core/api/ApiResponse"

export interface FilterWithPaginationCallTrackerRequest {
    PageSize: number
    PageNumber: number
    ProjectId?: number
    MobileNumber?: number
    Name?: string
    FromDate?: number
    ToDate?: number
    SortBy?: string
    ExportType?: 'Excel' | 'PDF'
}

export interface CallTrackerData {
    ProjectId: number | 0
    MobileNumber: number | 0
    Name: string | null
    FromDate: number | 0
    ToDate: number | 0
    CreatedById: number | 0
    CreatedBy: string | ''
    CreatedDate: string | null
    ModifiedById: number | 0
    ModifiedBy: string | ''
    ModifiedDate: string | null
    LastModifiedBy: string | ''
    LastModifiedDate: string | null
}

export type CallTrackerListResponse=ApiResponse<CallTrackerData[]>;